[PPPSYS-43233] examples moved up of the url sections
