+++ example/app/SiteSearch/page.tsx
