--- example/app/examples/page.tsx
@@ -1,5 +1,0 @@
-export default function Examples() {
-  return (
-    <div>examples</div>
-  )
-}
